Move test schemas to separate file
This way, they can be imported in all tests
